Fix --deep-black
The problem was that on perfectly vertical sides, one step would see
the cutter in contact with 100% black, and it would round it up to
the bottom level of the heightmap, and the next step would see
the cutter on the top surface of the part.

There wouldn't be any step where the cutter dipped below the heightmap
to remove the radius at the bottom of the corner.

The fix is to just always cut below the heightmap.